requirements.txt
Django==1.6.5
MySQL-python==1.2.5
Pillow==2.7.0
PyYAML==3.11
South==1.0
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.19
celery==3.1.17
django-celery==3.1.16
django-cors-headers==0.13
django-ipware==1.0.0
django-rest-swagger==0.3.1
djangorestframework==2.3.14
kombu==3.0.24
pytz==2014.10
qiniu==7.0.2
qrcode==5.1
requests==2.5.1
six==1.9.0
wsgiref==0.1.2
